+++ src/rendering/renderElement.js
@@ -9,9 +9,7 @@ export function renderElement(canvasElement) {
 
   const imageContext = getElementImageContext(canvasElement);
 
-  render(imageContext.imageData, imageContext.slice, imageContext.voiWindow, imageContext.canvasImageData);
-
-  imageContext.canvasContext.putImageData(imageContext.canvasImageData, 0, 0);
+  render(imageContext);
 
   canvasElement.dispatchEvent(new Event('ostrich.rendered', {
     bubbles: true,